+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Changelog
     subdirectory, maintaining backwards compatibility (@karalekas, gh-1143).
 -   Added a `.travis.yml` file to enable Travis CI for external-contributor builds,
     and upgraded GitLab CI style checks to py37 (@karalekas, gh-1145).
+-   Added typing to the `pyquil/experiment` module and added the module to the
+    `check-types` CI job (@karalekas, gh-1146).
 
 ### Bugfixes